Non successful tries for improved anchor smoothing

+++ video_processing/inference/optimization/annotate_tracklets.py
@@ -77,7 +77,7 @@ def _build_metadata(tracks: list[Track], input_path: Path, video_props: VideoInf
                         interpolated=det.interpolated,
                         boom=[float(det.boom.point.x), float(det.boom.point.y), float(det.boom.conf)],
                         mast_tip=[float(det.mast_tip.point.x), float(det.mast_tip.point.y), float(det.mast_tip.conf)],
-                        anchor=[int(det.anchor.x), int(det.anchor.y)],
+                        anchor=[float(det.anchor.x), float(det.anchor.y)],
                         scale=float(det.scale),
                     )
                     for det in rt.sorted_detections
